move psnr & ssim
