+++ torchmetrics/image/psnr.py
