+++ torchmetrics/image/ssim.py
